+++ src/plugin/jwt/index.js
@@ -63,32 +63,26 @@ class JwtPlugin extends BasePlugin {
       return res;
     }
 
+    const config = plugin.config.config;
     const creds = plugin.server.utils.parse_bearer_authorization_header(
       req.headers.authorization
     );
 
-    let token;
-    let valid = plugin.config.configs.some(config => {
-      try {
-        token = jwt.verify(creds.token, config.secret, config.options);
-        return true;
-      } catch (e) {
-        switch (e.name) {
-          case "TokenExpiredError":
-          case "JsonWebTokenError":
-          case "NotBeforeError":
-            error = e.name;
-            error_description = e.message;
-            break;
-          default:
-            throw e;
-        }
-      }
-    });
-
-    if (valid) {
+    try {
+      const token = jwt.verify(creds.token, config.secret, config.options);
       res.statusCode = 200;
       return res;
+    } catch (e) {
+      switch (e.name) {
+        case "TokenExpiredError":
+        case "JsonWebTokenError":
+        case "NotBeforeError":
+          error = e.name;
+          error_description = e.message;
+          break;
+        default:
+          throw e;
+      }
     }
 
     failure_response();