+++ CHANGELOG.md
@@ -80,3 +80,5 @@
 - ~~ensure all features are documented~~
 - ~~document limitations when service provier only allows 1 active token per client_id~~
 - ~~activity based session expiry (floating window sessions)~~
+- jwt assertions
+- ~~forward auth plugin~~